--- src-tauri/src/types.rs
@@ -1,5 +1,0 @@
-use crate::utils::workspace::Workspace;
-
-pub struct AppState {
-    pub active_workspace: Option<Workspace>,
-}
