+++ src-tauri/src/types/mod.rs
@@ -1,0 +1,2 @@
+mod types;
+pub use types::*;
