+++ src-tauri/src/workspace/mod.rs
@@ -1,0 +1,2 @@
+mod workspace;
+pub use workspace::*;
